[AD] Remove ECSFargateListener in favor of ContainerListener
We have discussed this internally a few times, after noticing that the
behavior of these two listeners is equivalent. To reduce code
duplication, we're removing ECSFargateListener completely, mapping any
explicit configuration to use ContainerListener, and having it always
run in ECS Fargate.
