+++ pkg/autodiscovery/listeners/container.go
@@ -9,6 +9,8 @@
 package listeners
 
 import (
+	"encoding/json"
+	"fmt"
 	"sort"
 	"strings"
 	"time"
@@ -18,6 +20,12 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/docker"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/workloadmeta"
+)
+
+const (
+	newIdentifierLabel            = "com.datadoghq.ad.check.id"
+	legacyIdentifierLabel         = "com.datadoghq.sd.check.id"
+	containerADTemplateCheckNames = "com.datadoghq.ad.check_names"
 )
 
 func init() {
@@ -75,6 +83,10 @@ func (l *ContainerListener) createContainerService(entity workloadmeta.Entity) {
 		}
 	}
 
+	if !container.State.Running && container.Runtime == workloadmeta.ContainerRuntimeECSFargate {
+		return
+	}
+
 	ports := make([]ContainerPort, 0, len(container.Ports))
 	for _, port := range container.Ports {
 		ports = append(ports, ContainerPort{
@@ -89,7 +101,7 @@ func (l *ContainerListener) createContainerService(entity workloadmeta.Entity) {
 
 	svc := &service{
 		entity: container,
-		adIdentifiers: ComputeContainerServiceIDs(
+		adIdentifiers: computeContainerServiceIDs(
 			containers.BuildEntityName(string(container.Runtime), container.ID),
 			containerImg.RawName,
 			container.Labels,
@@ -159,4 +171,48 @@ func findKubernetesInLabels(labels map[string]string) bool {
 		}
 	}
 	return false
+}
+
+// computeContainerServiceIDs takes an entity name, an image (resolved to an
+// actual name) and labels and computes the service IDs for this container
+// service.
+func computeContainerServiceIDs(entity string, image string, labels map[string]string) []string {
+	// ID override label
+	if l, found := labels[newIdentifierLabel]; found {
+		return []string{l}
+	}
+	if l, found := labels[legacyIdentifierLabel]; found {
+		log.Warnf("found legacy %s label for %s, please use the new name %s",
+			legacyIdentifierLabel, entity, newIdentifierLabel)
+		return []string{l}
+	}
+
+	ids := []string{entity}
+
+	// Add Image names (long then short if different)
+	long, short, _, err := containers.SplitImageName(image)
+	if err != nil {
+		log.Warnf("error while spliting image name: %s", err)
+	}
+	if len(long) > 0 {
+		ids = append(ids, long)
+	}
+	if len(short) > 0 && short != long {
+		ids = append(ids, short)
+	}
+	return ids
+}
+
+// getCheckNamesFromLabels unmarshals the json string of check names defined in
+// container labels and returns a slice of check names
+func getCheckNamesFromLabels(labels map[string]string) ([]string, error) {
+	if checkLabels, found := labels[containerADTemplateCheckNames]; found {
+		checkNames := []string{}
+		err := json.Unmarshal([]byte(checkLabels), &checkNames)
+		if err != nil {
+			return nil, fmt.Errorf("Cannot parse check names: %v", err)
+		}
+		return checkNames, nil
+	}
+	return nil, nil
 }
